Provide an alternative structure of Maze Game
The aim is to demonstrate a value of the Factory Method pattern.
The Game trait (interface) should be decoupled modularly from impls.
It would demonstrate that there could be many implementations, and
the game runs similarly with different ones.
